Extract resume picker loading
